+++ cmsafvis/R/render_plot_side_by_side.R
@@ -163,21 +163,21 @@ render_plot_side_by_side <- function(plot_rinstat,
           
           kml_toolbox <- raster::rasterToPolygons(ras)
           
-          # check_package_dependency("plotKML", reason = "exporting KML files")
-          # 
-          # plotKML::plotKML(
-          #   kml_toolbox,
-          #   file = outfile,
-          #   kmz = FALSE,
-          #   open.kml = FALSE,
-          #   plot.labpt = FALSE,
-          #   overwrite = TRUE,
-          #   outline = 0
-          # )
+          check_package_dependency("plotKML", reason = "exporting KML files")
+
+          plotKML::plotKML(
+            kml_toolbox,
+            file = outfile,
+            kmz = FALSE,
+            open.kml = FALSE,
+            plot.labpt = FALSE,
+            overwrite = TRUE,
+            outline = 0
+          )
            
-          cat("Sorry, but the plotKML R-package was removed from CRAN 
-              and KML output is not possible at the moment.
-              We are working on a solution for the next update.","\n")
+          # cat("Sorry, but the plotKML R-package was removed from CRAN 
+          #     and KML output is not possible at the moment.
+          #     We are working on a solution for the next update.","\n")
           
         } else if (fileExtension == ".tif") {
           dta <- as.vector(visualizeDataTimestep)
@@ -785,21 +785,21 @@ render_plot_side_by_side <- function(plot_rinstat,
           
           kml_toolbox <- raster::rasterToPolygons(ras)
           
-          # check_package_dependency("plotKML", reason = "exporting KML files")
-          # 
-          # plotKML::plotKML(
-          #   kml_toolbox,
-          #   file = outfile,
-          #   kmz = FALSE,
-          #   open.kml = FALSE,
-          #   plot.labpt = FALSE,
-          #   overwrite = TRUE,
-          #   outline = 0
-          # )
+          check_package_dependency("plotKML", reason = "exporting KML files")
+
+          plotKML::plotKML(
+            kml_toolbox,
+            file = outfile,
+            kmz = FALSE,
+            open.kml = FALSE,
+            plot.labpt = FALSE,
+            overwrite = TRUE,
+            outline = 0
+          )
           
-          cat("Sorry, but the plotKML R-package was removed from CRAN 
-              and KML output is not possible at the moment.
-              We are working on a solution for the next update.","\n")
+          # cat("Sorry, but the plotKML R-package was removed from CRAN 
+          #     and KML output is not possible at the moment.
+          #     We are working on a solution for the next update.","\n")
           
         } else if (fileExtension == ".tif") {
           dta <- as.vector(visualizeDataTimestep)